rename `index.ts` to `parser.ts`
