+++ lib/combinators/alternation/any.ts
@@ -1,4 +1,4 @@
-import { createParser, type Parser } from "../../../src/index.ts";
+import { createParser, type Parser } from "../../../src/parser.ts";
 import { sortPosition } from "../../../src/utilities.ts";
 
 /**
