+++ lib/primitives/take.ts
@@ -1,5 +1,5 @@
 import { parseErrors } from "../../src/errors.ts";
-import { createParser, type Parser } from "../../src/index.ts";
+import { createParser, type Parser } from "../../src/parser.ts";
 import { updatePosition } from "../../src/utilities.ts";
 
 /**
